+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # CHANGELOG
 
+## Version 2.6.0
+* Block CSI Driver (VCSI-193)
+
 ## Version 2.5.0
 * CSI driver operator (VCSI-173)
 * Allow using VIPPool DNS name instead of the CSI choosing IPs (VCSI-167)
 * Expose existing data via Static PV (VCSI-150)
+
+## Version 2.4.3
+* Support for multiple clusters via a single global secret
 
 ## Version 2.4.2
 * Support for ARM architecture (VCSI-191)